move early-game stuff
also i added lightningcss cause it's mildly faster, also turned *-build.js files into es modules
