+++ package.json
@@ -33,10 +33,12 @@
     "@stylistic/eslint-plugin": "^5.4.0",
     "@types/bun": "^1.3.0",
     "@vitejs/plugin-vue2": "^2.3.4",
+    "browserslist": "^4.26.3",
     "browserslist-useragent-regexp": "^3.0.2",
     "eslint": "^9.37.0",
     "eslint-plugin-oxlint": "^1.23.0",
     "eslint-plugin-vue": "^10.5.0",
+    "lightningcss": "^1.30.2",
     "oxlint": "^1.23.0",
     "rolldown-vite": "^7.1.17",
     "stylelint": "^16.25.0",
