+++ src/core/early-game/dimboost.js
@@ -1,5 +1,5 @@
-import { DC } from "./constants";
-import { EternityChallenge } from "./eternity/eternity-challenge";
+import { DC } from "../constants";
+import { EternityChallenge } from "../eternity/eternity-challenge";
 
 class DimBoostRequirement {
   constructor(tier, amount) {
