+++ src/core/early-game/galaxy.js
@@ -1,4 +1,4 @@
-import { DC } from "./constants";
+import { DC } from "../constants";
 
 export const GALAXY_TYPE = {
   NORMAL: 0,
