+++ src/core/early-game/index.js
@@ -1,0 +1,6 @@
+export * from "./dimboost.js";
+export * from "./sacrifice.js";
+export * from "./tickspeed.js";
+export * from "./tutorial.js";
+export * from "./galaxy.js";
+export * from "./normal-challenges.js";
