+++ src/core/early-game/sacrifice.js
@@ -1,4 +1,4 @@
-import { DC } from "./constants";
+import { DC } from "../constants";
 
 export class Sacrifice {
   // This is tied to the "buying an 8th dimension" achievement in order to hide it from new players before they reach
