+++ src/core/early-game/tutorial.js
